fix: move multisig error types into aa-accounts, add exports
