+++ packages/accounts/src/msca/plugins/multisig/index.ts
@@ -13,7 +13,9 @@ export {
 } from "./plugin.js";
 export type * from "./types.js";
 export { type SignerType } from "./types.js";
-export { combineSignatures } from "./utils/combineSignatures.js";
-export { formatSignatures } from "./utils/formatSignatures.js";
-export { getSignerType } from "./utils/getSignerType.js";
-export { splitAggregatedSignature } from "./utils/splitAggregatedSignature.js";
+export {
+  combineSignatures,
+  formatSignatures,
+  getSignerType,
+  splitAggregatedSignature,
+} from "./utils/index.js";
