--- packages/core/src/errors/types.ts
@@ -1,8 +1,0 @@
-import { BaseError } from "./base.js";
-
-export class InvalidContextSignatureError extends BaseError {
-  override name = "InvalidContextSignatureError";
-  constructor() {
-    super("Expected context.signature to be a hex string");
-  }
-}
